Suche nach Refinements für verschiedene Attributtypen optimieren

+++ cpp/subprojects/common/include/mlrl/common/model/body_conjunctive.hpp
@@ -99,7 +99,7 @@ class MLRLCOMMON_API ConjunctiveBody final : public IBody {
                  * @param threshold     The threshold
                  * @return              True, if the feature value satisfies the threshold, false otherwise
                  */
-                inline bool operator()(const float32& featureValue, const float32& threshold) const {
+                inline bool operator()(const int32& featureValue, const int32& threshold) const {
                     return featureValue <= threshold;
                 }
         };
@@ -117,7 +117,7 @@ class MLRLCOMMON_API ConjunctiveBody final : public IBody {
                  * @param threshold     The threshold
                  * @return              True, if the feature value satisfies the threshold, false otherwise
                  */
-                inline bool operator()(const float32& featureValue, const float32& threshold) const {
+                inline bool operator()(const int32& featureValue, const int32& threshold) const {
                     return featureValue > threshold;
                 }
         };
@@ -135,7 +135,7 @@ class MLRLCOMMON_API ConjunctiveBody final : public IBody {
                  * @param threshold     The threshold
                  * @return              True, if the feature value satisfies the threshold, false otherwise
                  */
-                inline bool operator()(const float32& featureValue, const float32& threshold) const {
+                inline bool operator()(const int32& featureValue, const int32& threshold) const {
                     return featureValue == threshold;
                 }
         };
@@ -153,7 +153,7 @@ class MLRLCOMMON_API ConjunctiveBody final : public IBody {
                  * @param threshold     The threshold
                  * @return              True, if the feature value satisfies the threshold, false otherwise
                  */
-                inline bool operator()(const float32& featureValue, const float32& threshold) const {
+                inline bool operator()(const int32& featureValue, const int32& threshold) const {
                     return featureValue != threshold;
                 }
         };
@@ -162,13 +162,13 @@ class MLRLCOMMON_API ConjunctiveBody final : public IBody {
 
         ConditionVector<float32, CompareNumericalGr> numericalGrVector_;
 
-        ConditionVector<float32, CompareOrdinalLeq> ordinalLeqVector_;
+        ConditionVector<int32, CompareOrdinalLeq> ordinalLeqVector_;
 
-        ConditionVector<float32, CompareOrdinalGr> ordinalGrVector_;
+        ConditionVector<int32, CompareOrdinalGr> ordinalGrVector_;
 
-        ConditionVector<float32, CompareNominalEq> nominalEqVector_;
+        ConditionVector<int32, CompareNominalEq> nominalEqVector_;
 
-        ConditionVector<float32, CompareNominalNeq> nominalNeqVector_;
+        ConditionVector<int32, CompareNominalNeq> nominalNeqVector_;
 
     public:
 
@@ -199,23 +199,23 @@ class MLRLCOMMON_API ConjunctiveBody final : public IBody {
          * An iterator that provides access to the thresholds that are used by ordinal conditions in the body and allows
          * to modify them.
          */
-        typedef View<float32>::iterator ordinal_threshold_iterator;
+        typedef View<int32>::iterator ordinal_threshold_iterator;
 
         /**
          * An iterator that provides read-only access to the thresholds that are used by ordinal conditions in the body.
          */
-        typedef View<float32>::const_iterator ordinal_threshold_const_iterator;
+        typedef View<int32>::const_iterator ordinal_threshold_const_iterator;
 
         /**
          * An iterator that provides access to the threshold that are used by nominal conditions in the body and allows
          * to modify them.
          */
-        typedef View<float32>::iterator nominal_threshold_iterator;
+        typedef View<int32>::iterator nominal_threshold_iterator;
 
         /**
          * An iterator that provides read-only access to the thresholds that are used by nominal conditions in the body.
          */
-        typedef View<float32>::const_iterator nominal_threshold_const_iterator;
+        typedef View<int32>::const_iterator nominal_threshold_const_iterator;
 
         /**
          * An iterator that provides access to the feature indices that correspond to the conditions in the body and
